Refactored dto structure to reside in a single package
