+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/AccountController.kt
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import pt.up.fe.ni.website.backend.model.dto.AccountDto
+import pt.up.fe.ni.website.backend.dto.entity.AccountDto
 import pt.up.fe.ni.website.backend.service.AccountService
 
 @RestController
