+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/AuthController.kt
@@ -6,16 +6,9 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import pt.up.fe.ni.website.backend.dto.auth.LoginDto
+import pt.up.fe.ni.website.backend.dto.auth.TokenDto
 import pt.up.fe.ni.website.backend.service.AuthService
-
-data class LoginDto(
-    val email: String,
-    val password: String
-)
-
-data class TokenDto(
-    val token: String
-)
 
 @RestController
 @RequestMapping("/auth")
