+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/PostController.kt
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import pt.up.fe.ni.website.backend.model.dto.PostDto
+import pt.up.fe.ni.website.backend.dto.entity.PostDto
 import pt.up.fe.ni.website.backend.service.PostService
 
 @RestController
