+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/ProjectController.kt
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import pt.up.fe.ni.website.backend.model.dto.ProjectDto
+import pt.up.fe.ni.website.backend.dto.entity.ProjectDto
 import pt.up.fe.ni.website.backend.service.ProjectService
 
 @RestController
