+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/auth/LoginDto.kt
@@ -1,0 +1,6 @@
+package pt.up.fe.ni.website.backend.dto.auth
+
+data class LoginDto(
+    val email: String,
+    val password: String
+)
