+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/auth/TokenDto.kt
@@ -1,0 +1,5 @@
+package pt.up.fe.ni.website.backend.dto.auth
+
+data class TokenDto(
+    val token: String
+)
