+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/entity/EventDto.kt
@@ -1,4 +1,4 @@
-package pt.up.fe.ni.website.backend.model.dto
+package pt.up.fe.ni.website.backend.dto.entity
 
 import pt.up.fe.ni.website.backend.model.Event
 import java.util.Date
@@ -7,4 +7,4 @@ class EventDto(
     val title: String,
     val description: String,
     val date: Date
-) : Dto<Event>()
+) : EntityDto<Event>()
