+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/entity/PostDto.kt
@@ -1,4 +1,4 @@
-package pt.up.fe.ni.website.backend.model.dto
+package pt.up.fe.ni.website.backend.dto.entity
 
 import pt.up.fe.ni.website.backend.model.Post
 
@@ -6,4 +6,4 @@ class PostDto(
     val title: String,
     val body: String,
     val thumbnailPath: String
-) : Dto<Post>()
+) : EntityDto<Post>()
