+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/entity/ProjectDto.kt
@@ -1,4 +1,4 @@
-package pt.up.fe.ni.website.backend.model.dto
+package pt.up.fe.ni.website.backend.dto.entity
 
 import pt.up.fe.ni.website.backend.model.Project
 
@@ -6,4 +6,4 @@ class ProjectDto(
     val title: String,
     val description: String,
     val isArchived: Boolean = false
-) : Dto<Project>()
+) : EntityDto<Project>()
