+++ src/test/kotlin/pt/up/fe/ni/website/backend/controller/AuthControllerTest.kt
@@ -17,6 +17,8 @@ import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.get
 import org.springframework.test.web.servlet.post
+import pt.up.fe.ni.website.backend.dto.auth.LoginDto
+import pt.up.fe.ni.website.backend.dto.auth.TokenDto
 import pt.up.fe.ni.website.backend.model.Account
 import pt.up.fe.ni.website.backend.model.CustomWebsite
 import pt.up.fe.ni.website.backend.repository.AccountRepository
